requirements.txt
# Python dependencies for Club Analytics Pipeline

# Core data processing
pandas>=1.5.0
numpy>=1.24.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Visualization
streamlit>=1.28.0
plotly>=5.17.0

# Data quality and testing
pytest>=7.4.0
pytest-cov>=4.1.0
soda-core-postgres>=3.0.0

# dbt
dbt-postgres>=1.6.0

# Airflow (for local development)
apache-airflow>=2.7.0
apache-airflow-providers-postgres>=5.0.0

# Development tools
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.5.0

# Additional utilities
python-dotenv>=1.0.0
requests>=2.31.0


